modified:   app/(user)/dashboard/page.tsx 	modified:   app/actions/add-money.ts 	modified:   app/actions/auth/log-out.ts 	modified:   app/actions/auth/login.ts 	modified:   app/actions/auth/sign-up.ts 	modified:   app/actions/delete-money.ts 	modified:   app/actions/edit-money.ts 	modified:   app/actions/get-moneys.ts 	modified:   app/actions/get-session.ts 	modified:   components/forms/log-in.tsx 	modified:   components/forms/sign-up.tsx 	modified:   components/nav.tsx 	modified:   components/side-nav.tsx

+++ app/actions/delete-money.ts
@@ -1,12 +1,15 @@
 "use server";
 import { spServer } from "@/lib/supabase/server";
 import { cookies } from "next/headers";
+import { redirect } from "next/navigation";
 
 export const delmoney = async (id: string) => {
   const supabase = spServer(cookies());
 
-  const { error } = await supabase.from("moneys").delete().eq("id", id);
+  const { data: session } = await supabase.auth.getSession();
+  if (!session.session) return redirect("/log-in");
 
+  const { error } = await supabase.from("moneys").delete().eq("id", id);
   if (error) return { error: error };
   return { success: "ok" };
 };
